+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(solana-utils): Added `luts` and `luts_mut` methods for `BundleBuilder`.
 - sdk(sdk): Added `SerdeMarketConfigBuffer`.
 - sdk(decode): Added `TransactionAccess` trait to abstract transaction decoding logic.
+- sdk(sdk): Added `MarketClosed` error type.
 - cli: Added `market buffer` command to display the content of given buffer account.
 - cli: Added TOML as output format.
 - cli: Added output of ALT address for `alt extend` command.